+++ _archive/FilteredSampleProd.py
@@ -1,9 +1,9 @@
 {
-  "id": 40,
-  "sku": "Cup-2",
+  "id": 400,
+  "sku": "TT-01",
   "ebay_account_id": 1,
-  "title": "Copeland Spode Queen Elizabeth II 1953 Coronation Mug Tankard",
-  "description": "This is a Copeland Spode mug commemorating the coronation of Queen Elizabeth II on June 2nd, 1953. The mug has a creamy white glaze with brown transfer printing and coloured crests. It features a portrait of Queen Elizabeth II, the royal crest, and text marking the occasion.",
+  "title": "Copeland Spode Queen Elizabeth II 1953 Coronation Cup and Saucer",
+  "description": "This is a Copeland Spode cup and saucer commemorating the coronation of Queen Elizabeth II on June 2nd, 1953. The mug has a creamy white glaze with brown transfer printing and coloured crests. It features a portrait of Queen Elizabeth II, the royal crest, and text marking the occasion.",
   "condition_description": "Used condition with staining and minor chips around the rims. The glaze is crazed and discolored. There are marks and scuffs on the exterior surface.",
   "suggested_price": 12,
   "selected_category_path": "Historical Memorabilia >> Collectables >> Royalty Collectables",